Issue #43 - Port oco3.val module tests to bliss-core
- Port over oco3.val unit tests to bliss-core. There is now some
  duplication of test coverage with the previously added YAMLProcessor
  class tests, but for the most part this is fairly minor.

- Move oco3.val test data over to bliss-core. There are now a number of
  (in)valid CMD and TLM yaml files in bliss/test/testdata for use in
  test validation.

- Move oco3 cmd, evr, and tlm schemas to bliss example config data.

Note, it's possible that much of the tlm/cmd YAML files and example
schema information will need to be cleaned up a bit since these were
pulled from oco3mos. At the moment this give us a decent chunk of data
to exercise the bliss.val module with so they're being included.