Issue #189 - Normalize table constructor format
table.yaml now uses the same constructor formating found in cmd and tlm
yaml files. This requires an update to the way that the table.yaml is
loaded.

The table module now handles multiple documents with nested FSWTables
when loading table.yaml. This allows it handle the new table yaml format
loads. This is not backwards compatible with the previous format. This
also fixes a bug in the FSWTabDict.load function in which no file would
be loaded if FSWTabDict was instantiated without a filename originally.

For example, this used to not work:

t = bliss.table.FSWTabDict()
t.load('/path/to/table.yaml')

This was due to incorrect nesting of the code in the load function. None
of the code was executed if the FSWTabDict's filename attribute was None
(which would be the case in the above example) even though the
instantiation of the original class instance wasn't invalid.