Issue #196 - Update bliss-core EVR format
bliss.core now supports an expanded EVR format that is similar to the
existing tlm and cmd YAML config formats. An EVR is now defined as

- !EVR
  name: NO_ERROR
  code: 0x0001
  desc: No error
  message: "No error %s"

The 'message' attribute can contain printf formatters and support for
automatic unpacking of the formatted values will be added in follow on
tickets.

Note, the util.toDict function doesn't generate the nicest dictionary
format during testing. It ends up with oddly nested objects that don't
match the expected result. As such, the new EVRDefn object current
manually handles conversion to a dict, but it maintains a consistent
interface. The util.toDict oddities will be addressed in another ticket.