Issue #196 - Update dtype EVRType for new EVR format handling
dtype EVRType encoding has been updated so that it properly searches for
a given EDRDefn object given an EVR name. Encoding call now expects the
EVR code insted of a tuple. The dtype decoding decode has to search the
EVR definitions for a matching EVR code instead of the previous
dict-index due to the new format.

dtype testing as been updated with the new EVRDefn formatting as well.