Issue #234 - Update loading of schema files for YAML validation
Schema files have been moved to bliss/core/data so they can be loaded as
package data and setup.py has been updated include those files. The tlm,
cmd, and evr functions for accessing schema files have been updated to
use the pkg_resources approach to locating the schema files. Tests have
been updated to use the API function if possible (EVR wasn't) and the
table test has been updated to use pkg_resources since it doesn't seem
to provide an interface for this.

MANIFEST.in has been dropped. None of the files included were being
properly moved as package_data / data_files. If we decide we need it for
something in the future we can add it back in.