Issue #293 - Add contributor guide to docs
Add a contributor to Sphinx documentation. All developer / contributor
information has been removed from the README and placed into the
contributor guide if it was still needed. The README has been updated
with a link to the new contributor documentation. Note, this link will
not work until the docs are built for the next release.

README layout has been rearranged slightly.