Added exceptions APIError and APITimeoutError; updated wait()
All BLISS API exceptions are derived from APIError.  The
APITimeoutError exception is raised when a timeout limit is exceeded.

The wait() method now takes a `_timeout` parameter (was `timeout`) and
a `_raiseException` parameter:

Waits either a specified number of seconds, e.g.:

    wait(1.2)

or for a given condition to be True.  Conditions may be take
several forms: Python string expression, lambda, or function,
e.g.:

    wait('instrument_mode == "SAFE"')
    wait(lambda: instrument_mode == "SAFE")

    def isSafe(): return instrument_mode == "SAFE"
    wait(isSafe)

The default ``_timeout`` is 10 seconds.  If the condition is not
satisfied before the timeout has elapsed, an
:exception:``APITimeoutError`` exception is raised.

The :exception:``APITimeoutError`` exception may be supressed in favor
of returning ``True`` on success (i.e. condition satisfied) and
``False`` on failure (i.e. timeout exceeded) by setting the
``_raiseException`` parameter to ``False``.

These parameters are prefixed with an underscore so they may also be
used to control exception handling when sending commands.  Since
methods that generate commands take keyword arguments, we did not want
these parameter names to conflict with command parameter names.