TIME32, 40, and 64 now encode and decode Python datetime objects
TIME32, TIME40, and TIME64 currently encode() based on a string
representation of date and time. It's better to encode() from (and
decode() to) a Python datetime and let users choose their own string
representation. There's one issue here: Python datetime has only
microsecond resolution, but a TIME64 has nanosecond resolution. If
this becomes a problem for our users, we could subclass datetime and
store the nanoseconds explicitly.

The encode() methods raise a Python TypeError if not passed a Python
datetime.

This commit also removes:

    self._name = [name for name in ComplexTypeNames.keys()
                     if ComplexTypeNames[name] == self.__class__][0]

to avoid a dependence on ComplexTypeNames.  If anything, the
ComplexTypeNames map should be constructed from the name of the type,
not the other way around.