Issue #359 - Packet CMD, EVR, and TIME support raw decodes
Resolves #359.

BLISS allows the raw value of a packet field to be retrieved with the
following syntax:

    packet.raw.fieldname

But this worked only with enumerated value fields and fields where a
DN-to-EU conversion (`dntoeu:` YAML block) would be applied.  This
commit extends this behavior to:

  * Commands (return the raw opcode, not a BLISS `CmdDefn`),
  * EVRs     (return the raw opcode, not a BLISS `EVRDefn`),
  * Times    (return the raw value,  not a Python `datetime`)

As well as:

  * `ArrayType`s
  * [LSB|MSB]_[F32|D64]

In the case of floating-point types (e.g. `MSB_F32`) raw decodes will
return the underlying bytes representing the floating-point number as
an unsigned integer (i.e. `U32` or `U64`).

This commit also adds unit tests and cleans up some unnecessary
`bliss.dtype` module level variables.