Issue #364 - Update cfg path validity check to handle list of paths
Config path validation now handles the case where the result of
expanding the base path results in a list of paths instead of just a
single path value. This is most likely to happen in the data path
expansion after variable substitution.

The warning has been updated to include information on the path that
failed to verify. This ensure that in the case where one key maps to a
list of paths that the warning message specifies which of the paths
didn't validate, thus avoiding duplicate error messages that don't help
the user debug the problem.