Issue #377 - Raw F32 or D64 access should yield a float
Resolves #377.

BLISS allows the raw value of a packet field to be retrieved with the
following syntax:

    packet.raw.fieldname

Commit 73d93ae6788b8f1a406b31ef93c389be8c868faa / Issue #359
introduced a "feature" where if the raw access to the field in
question is an `[LSB|MSB]_[F32|D64]` it is decoded not as a
floating-point number, but the corresponding integer value, i.e. a
`[LSB|MSB]_U[32|64]`.  This wasn't very well thought.  It breaks
DN-to-EU conversions for Cold Atom Lab (CAL).

For example:

```
 constants:
   A: 42.9

 functions:
   PMON(dn): (dn/A)

- !Field
  name:   Pd1
  desc:
  type:   MSB_D64
  dntoeu:
    equation: PMON(raw.Pd1)
    units: milliwatts
```

So, it seems prudent to revert this change such that, for
floating-point primitive types `packet.raw.fieldname` once again
returns a floating-point value.