Issue #397: refactor val.ErrorHandler to better leverage JSON Path
JSON Schema returns a JSON relative path to track down the specific
location of an error in a file. Leveraging this JSON path, we are
able to loop through and find specific errors thrown by the jsonschema
validator.

val.py
-------
* refactor ErrorHandle class to better leverage the JSON path
* for instance, looking for the 50th desc: versus the first/last like in the past. this was causing a lot of problems in the specific case, because this issue was identical throughout the dictionary so there is no easy way to track it down using the key values.
* this will make it a lot easier to track down issues that arise from schema validation

test_val.py and testTlmValidator6.yaml
--------------------------------------------
new unit test to try this use case where desc: is an empty string