Issue #416: update tlm.py to handle single-item lists
this update will allow tlm.py more gracefully handle if
someone specifies a single-item list instead of an integer
for a U8 tlm field. the update simply checks the bytes in the
setter and translates to an integer if it is a single-item list.

there is also a test to make sure we can properly handle this.