Issue #454 - Broken: bliss-pcap --stats via pcap.stats()
This commit fixes `pcap.stats()`, renames it to `pcap.times()`, and
updates the documentation string to use Python's standard reStructured
text.  The function `pcap.times()` now returns a dictionary keyed by
filename, with each value a list of (start, stop) time ranges for that
file.

This commit updates the `bliss-pcap` command-line utility, replacing
`--stats` with `--times` and upating it to call `pcap.stats()`.  If
more than one pcap file is processed, the pcap filename is prepended
to the time range in the `--times` output.

Finally, in following with the Unix convention, the pcap files to
process are now positional arguments and no longer require a `--pcap`
option.

Fixes #454.