Issue #465: persistence and when conditions added
These updates only include those to allow persistence and
when conditions to be specified in limits.yaml and be
accessible through the limits API.

There is no logic to engage these in the limit checks.

There are also no units tests since there is not really
any functionality to test.