Issue #31 - Add text and email notification support
Add `bliss.core.notify` for sending email and text notifications. The
`notify` module has support for sending email and text (or both)
notifications via SMTP.

Add new bliss.config section for notifications. Notification supports
config for email and text notifications and for SMTP server connection
information.

A notification type has a list of "triggers" which specifies the name(s)
of triggers for which that type of event should be emitted. These are
used in conjunction with `bliss.config.notify.trigger_notification()`
so that components can submit triggers for notifications while allowing
the user to control what type (if any) of notification is emitted.

A notification type also has a list of "recipients". For both email and
text notifications these are email addresses.

SMTP server information is configured via 4 parameters. All of these are
expected to be set. Un-auth'ed SMTP connections are not supported.

    notifications:
        email:
            triggers:
                - limit-warn
                - limit-error
            recipients:
                - email@address.com
        text:
            triggers:
                - limit-warn
                - limit-error
            recipients:
                - phonenumber@service.com
        smtp:
            server: smtp.gmail.com
            port: 465
            username: yourSMTPServerUsername
            password: yourSMTPServerPassword