GUI Issue #75 - Updates to stream input config and parsing
Now allow all streams to have multiple inputs, and set the inputs
attribute to an empty list instead of None if no inputs specified.

Moved input from PortInputClient init params to kwargs for consistency
across client subclasses.

PortInputClient will now look for a stream's first input to be a port
on which it listens for messages.