Issue #216 and #217 - Fix default plugins and handler package
Issue #216 - Update server.plugins default value in config to an empty
list so it doesn't fall back to None. This prevents it from breaking
other code that checks if the value exists and falls back to a sane
default if it doesn't.

Issue #217 - Handler class path wasn't updated as part of the changes in
issue #214.