Database API Result Handling and SQLite bug fixes and improvements
This is a bit of a sweeping rework to parts of the database APIs to
provide a unification to result formatting. During testing a few issues
were run into specific to the SQLiteBackend implementation that resulted
in a complete rework to the data storage format. These changes are
related enough that I'm grouping them under a single pull request.
Related issues:

Issue #303 - Database API query and results abstraction
Issue #307 - SQLite Backend Simplification
Issue #310 - DB API Sqlite connection and creation handling fails
    to properly cover cases

Add AITDBResult class to provide a minimal wrapper around all database
query results. Any API that executes a query and returns results now
returns an AITDBResult object. Users can access the queries execute, any
errors encountered, and the query results (either raw or as processed
Packets depending on the interface) through this object.

Add a `query_packets` function to the backends. These functions take an
optional list of packet types and start / end times and return
time-ordered Packet objects for all results. This is likely the primary
interface that users and AIT will use to access data in a
backend-agnostic way.

Add a `create_packet_from_result` classmethod to the backends. The
backends use these to parse a given query result item into an AIT
packet. Previously only the Influx backend had a similar function (it is
now marked deprecated). These also provide users a helpful interface
for converting results to packets as part of their custom queries.

Completely rework the SQLite backend storage format to use BLOB to store
packets instead of splitting fields out into individual columns.
Ticket #307 provides an explanation for why this was done. In short,
the SQLite support is really meant for demonstration purposes and to
provide a baseline implementation that other backends can reference.
Users are heavily encouraged to NOT use SQLite as their primary mission
backend. Now seemed like a good time to make this switch given the impact
on the existing code from changes for #303 and the lack of thorough testing
for the previous format and type handling.

Fix the poor SQLiteBackend handling of connect / create and fix the
issue where custom named databases weren't being properly configured.
SQLiteBackend.connect now "connects" to the datastore and checks if the
database previously existed. If not, then the table structure needs
handled. SQLiteBackend.create now only creates tables instead of the
previously incorrect hybrid of functionality.