Issue #339 - Document Packet / Field naming constraints
Update telemetry documentation with a warning that all Packet / Field
names should be valid Python variable names to avoid any issues when
using the toolkit.

Note, even though this can (sometimes) be worked around and it's not
necessarily always an issue in both the Packet and Field cases, I think
it's safer to just warn against any of this and avoid the overhead of
trying to debug various issues here.