Update Core package handling
Update Core to use Poetry for package management. See their home page
for additional details on how to use it and what it gets us.

https://python-poetry.org/

Update packaging to be PEP 517/518 compliant. All configuration has been
moved to setup.cfg and pyproject.toml. Setup.py has been dropped (since
we're not using setuptools now).

Documentation has been updated and moved from the wiki to the README.
Once we merge this we should drop the wiki page since a decent chunk of
that page is now out of date.

Resolve #364