Issue #459 - Update to support Python3.10
Because of the strictly pinned versions this library would only work
with Python 3.7, as stated in the documentation. Since the dependencies
are well founded they provide support for newer versions of Python,
including Python 3.10. By migrating to the newer versions this library
will now support Python3.10 but drop support for other versions.