Issue #376 - Update outdated sphinx-apidoc call in README
Update README's `sphinx-apidoc` call instructions to use
`--implicit-namespaces`. This is necessary to ensure that the `ait`
namespace is picked up and included in the API docs that are generated.
Without this files will be written starting with the `core` module.
E.g.:

    core.server.handlers.rst

    vs

    ait.core.server.handlers.rst

Update out-of-date API docs from previous server plugin additions.