Issue #467 - Add baseline GH Actions build
Add a default build for any push to master or pull request that runs
the full `tox` build. This includes the test suite for all supported
python versions, the docs build, and the linting pipeline.

This is sufficient for a first pass but there are a couple of things we
might want to poke at. I could not get `actions/setup-python` caching of
poetry dependencies to work with `tox-gh-actions`. The lint run kept
failing because it couldn't find `pre-commit` even though it seemed to
have been installed properly. Should poke at this in a follow on ticket.

Would also be good to get a better build running to test package
deployment and the like. Bundling everything up and shipping it off to
test.pypi would be great. I'd also like to get a release build
integrated so we aren't manually doing that. Automatic CHANGELOG updates
after a push to master or PR merge would also be great.

Resolve #467