Issue #491 - Move logger globals for Python bug
Shift the `log.py` logger globals to the beginning of the file to avoid
a Python bug that causes spurious "annotated name can't be global"
errors when mypy lints the files.

See here for the issue report:
https://bugs.python.org/issue34939

Clarification on the ticket marks this as a bug and a fix was added.
This avoids the error in Python 3.7.